requirements.txt
PyYAML==6.0.2
aioredis==2.0.1
fastapi==0.115.6
pydantic==2.10.5
kafka==1.3.5
